# docs/swagger.yaml
definitions:
...
  types.APIBadRequest:
    type: object
  types.APIInternalServerError:
    type: object
...
  types.Organization:
    properties:
      homepage:
        type: string
      logo:
        type: string
      name:
        type: string
      org_type:
        type: string
      path:
        description: unique name of the organization
        type: string
      user_id:
        type: integer
      verified:
        type: boolean
    type: object
...
  types.Response:
    properties:
      data: {}
      msg:
        type: string
    type: object
...
  types.UpdateUserRequest:
    properties:
      avatar:
        type: string
      bio:
        type: string
      email:
        type: string
      homepage:
        type: string
      name:
        description: Display name of the user
        type: string
      new_username:
        description: if use want to change username, this should be the only field
          to send in request body
        type: string
      phone:
        type: string
      roles:
        description: should be updated by admin
        example:
        - '[super_user'
        - ' admin'
        - ' personal_user]'
        items:
          type: string
        type: array
      uuid:
        type: string
    type: object
  types.User:
    properties:
      avatar:
        type: string
      bio:
        type: string
      can_change_username:
        type: boolean
      email:
        type: string
      homepage:
        type: string
      id:
        type: integer
      last_login_at:
        type: string
      nickname:
        type: string
      orgs:
        items:
          $ref: '#/definitions/types.Organization'
        type: array
      phone:
        type: string
      roles:
        items:
          type: string
        type: array
      username:
        type: string
      uuid:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /user/{username}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: username
        in: path
        name: username
        required: true
        type: string
      - description: current user
        in: query
        name: current_user
        required: true
        type: string
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/types.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.Response'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/types.APIBadRequest'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/types.APIInternalServerError'
      security:
      - ApiKey: []
      summary: Delete user
      tags:
      - User
    get:
      consumes:
      - application/json
      parameters:
      - description: username or uuid, defined by the query string 'type'
        in: path
        name: username
        required: true
        type: string
      - description: current user
        in: query
        name: current_user
        type: string
      - description: path param is usernam or uuid, default to username
        enum:
        - username
        - uuid
        in: query
        name: type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/types.Response'
            - properties:
                data:
                  $ref: '#/definitions/types.User'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/types.APIBadRequest'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/types.APIInternalServerError'
      security:
      - ApiKey: []
      summary: Get user info. Admin and the user self can see full info, other users
        can only see basic info.
      tags:
      - User
    put:
      consumes:
      - application/json
      description: update user
      parameters:
      - description: username
        in: path
        name: username
        required: true
        type: string
      - description: current user
        in: query
        name: current_user
        required: true
        type: string
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/types.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.Response'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/types.APIBadRequest'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/types.APIInternalServerError'
      security:
      - ApiKey: []
      summary: Update user. If change user name, should only send 'new_username' in
        the request body.
      tags:
      - User
...
  /users:
    get:
      consumes:
      - application/json
      parameters:
      - description: search
        in: query
        name: search
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/types.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/types.User'
                  type: array
                total:
                  type: integer
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/types.APIBadRequest'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/types.APIInternalServerError'
      security:
      - ApiKey: []
      summary: Get users info. Only Admin
      tags:
      - User
securityDefinitions:
  ApiKey:
    description: Bearer token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"